fix: move unit test to new file
